+++ ivy/array/activations.py
@@ -272,40 +272,3 @@ class ArrayWithActivations(abc.ABC):
         ivy.array([-0.30340147,  0.        ,  0.86509842])
         """
         return ivy.mish(self._data, out=out)
-
-    def thresholded_relu(
-        self: ivy.Array,
-        /,
-        *,
-        threshold: Optional[Union[int, float]] = 0,
-        out: Optional[ivy.Array] = None,
-    ) -> ivy.Array:
-        """
-        ivy.Array instance method variant of ivy.thresholded_relu.
-        This method simply wraps the function, and so the docstring
-        for ivy.thresholded_relu also applies to this method with minimal changes.
-
-        Parameters
-        ----------
-        self
-            input array.
-        threshold
-            threshold value above which the activation is linear. Default: ``0``.
-        out
-            optional output array, for writing the result to. It must have a shape
-            that the inputs broadcast to.
-
-        Returns
-        -------
-        ret
-            an array with the relu activation function applied element-wise
-            with custom threshold.
-
-        Examples
-        --------
-        >>> x = ivy.array([-1., .2, 1.])
-        >>> y = x.thresholded_relu(threshold=0.5)
-        >>> print(y)
-        ivy.array([0., 0., 1.])
-        """
-        return ivy.thresholded_relu(self._data, threshold=threshold, out=out)
