+++ ivy/container/activations.py
@@ -1040,133 +1040,3 @@ class ContainerWithActivations(ContainerBase):
             map_sequences=map_sequences,
             out=out,
         )
-
-    @staticmethod
-    def static_thresholded_relu(
-        x: Union[ivy.Array, ivy.NativeArray, ivy.Container],
-        /,
-        *,
-        threshold: Optional[Union[int, float]] = 0,
-        key_chains: Optional[Union[List[str], Dict[str, str]]] = None,
-        to_apply: bool = True,
-        prune_unapplied: bool = False,
-        map_sequences: bool = False,
-        out: Optional[ivy.Container] = None,
-    ) -> ivy.Container:
-        """
-        ivy.Container static method variant of ivy.thresholded_relu.
-        This method simply wraps the function, and so the docstring
-        for ivy.thresholded_relu also applies to this method with minimal changes.
-
-        Parameters
-        ----------
-        x
-            input container.
-        threshold
-            threshold value above which the activation is linear. Default: ``0``.
-        key_chains
-            The key-chains to apply or not apply the method to. Default is ``None``.
-        to_apply
-            If True, the method will be applied to key_chains, otherwise key_chains
-            will be skipped. Default is ``True``.
-        prune_unapplied
-            Whether to prune key_chains for which the function was not applied.
-            Default is ``False``.
-        map_sequences
-            Whether to also map method to sequences (lists, tuples).
-            Default is ``False``.
-        out
-            optional output container, for writing the result to. It must have a shape
-            that the inputs broadcast to.
-
-        Returns
-        -------
-        ret
-            a container with the rectified linear activation unit function
-            applied element-wise with custom threshold.
-
-        Examples
-        --------
-        >>> x = ivy.Container(a=ivy.array([1.0, -1.2]), b=ivy.array([0.4, -0.2]))
-        >>> y = ivy.Container.static_thresholded_relu(x, threshold=0.5)
-        >>> print(y)
-        {
-            a: ivy.array([1., 0.]),
-            b: ivy.array([0., 0.])
-        }
-
-        """
-        return ContainerBase.cont_multi_map_in_function(
-            "thresholded_relu",
-            x,
-            threshold=threshold,
-            key_chains=key_chains,
-            to_apply=to_apply,
-            prune_unapplied=prune_unapplied,
-            map_sequences=map_sequences,
-            out=out,
-        )
-
-    def thresholded_relu(
-        self: ivy.Container,
-        /,
-        *,
-        threshold: Optional[Union[int, float]] = 0,
-        key_chains: Optional[Union[List[str], Dict[str, str]]] = None,
-        to_apply: bool = True,
-        prune_unapplied: bool = False,
-        map_sequences: bool = False,
-        out: Optional[ivy.Container] = None,
-    ) -> ivy.Container:
-        """
-        ivy.Container instance method variant of ivy.thresholded_relu.
-        This method simply wraps the function, and so the docstring
-        for ivy.thresholded_relu also applies to this method with minimal changes.
-
-        Parameters
-        ----------
-        self
-            input container.
-        threshold
-            threshold value above which the activation is linear. Default: ``0``.
-        key_chains
-            The key-chains to apply or not apply the method to. Default is ``None``.
-        to_apply
-            If True, the method will be applied to key_chains, otherwise key_chains
-            will be skipped. Default is ``True``.
-        prune_unapplied
-            Whether to prune key_chains for which the function was not applied.
-            Default is ``False``.
-        map_sequences
-            Whether to also map method to sequences (lists, tuples).
-            Default is ``False``.
-        out
-            optional output container, for writing the result to. It must have a shape
-            that the inputs broadcast to.
-
-        Returns
-        -------
-        ret
-            a container with the rectified linear activation unit function
-            applied element-wise with custom threshold.
-
-        Examples
-        --------
-        >>> x = ivy.Container(a=ivy.array([1.0, -1.2]), b=ivy.array([0.4, -0.2]))
-        >>> y = x.thresholded_relu(threshold=0.5)
-        >>> print(y)
-        {
-            a: ivy.array([1., 0.]),
-            b: ivy.array([0., 0.])
-        }
-
-        """
-        return self.static_thresholded_relu(
-            self,
-            threshold=threshold,
-            key_chains=key_chains,
-            to_apply=to_apply,
-            prune_unapplied=prune_unapplied,
-            map_sequences=map_sequences,
-            out=out,
-        )
