+++ ivy/functional/backends/jax/experimental/activations.py
@@ -1,5 +1,7 @@
+from typing import Optional, Union
+
+# global
 import jax.numpy as jnp
-from typing import Optional
 from ivy.functional.backends.jax import JaxArray
 
 
@@ -9,3 +11,13 @@ def logit(x: JaxArray, /, *, eps: Optional[float] = None, out=None):
     else:
         x = jnp.clip(x, eps, 1 - eps)
     return jnp.log(x / (1 - x))
+
+
+def thresholded_relu(
+    x: JaxArray,
+    /,
+    *,
+    threshold: Optional[Union[int, float]] = 0,
+    out: Optional[JaxArray] = None,
+) -> JaxArray:
+    return jnp.where(x > threshold, x, 0).astype(x.dtype)
