+++ ivy/functional/backends/numpy/experimental/activations.py
@@ -1,5 +1,10 @@
+from typing import Optional, Union
+
+# global
 import numpy as np
-from typing import Optional
+
+# local
+from ivy.functional.backends.numpy.helpers import _scalar_output_to_0d_array
 
 
 def logit(x: np.ndarray, /, *, eps: Optional[float] = None, out=None):
@@ -12,3 +17,17 @@ def logit(x: np.ndarray, /, *, eps: Optional[float] = None, out=None):
     if np.isscalar(ret):
         return np.array(ret)
     return ret
+
+
+@_scalar_output_to_0d_array
+def thresholded_relu(
+    x: np.ndarray,
+    /,
+    *,
+    threshold: Optional[Union[int, float]] = 0,
+    out: Optional[np.ndarray] = None,
+) -> np.ndarray:
+    return np.where(x > threshold, x, 0).astype(x.dtype)
+
+
+thresholded_relu.support_native_out = True
