+++ ivy/functional/backends/tensorflow/experimental/__init__.py
@@ -1,3 +1,9 @@
+# global
+import tensorflow as tf
+
+backend_version = {"version": tf.__version__}
+
+
 # flake8: noqa
 from .activations import *
 from .creation import *
