+++ ivy/functional/backends/torch/experimental/__init__.py
@@ -1,3 +1,8 @@
+# global
+import torch as torch
+
+backend_version = {"version": torch.__version__.split("+")[0]}
+
 # flake8: noqa
 from .activations import *
 from .converters import *
