+++ ivy_tests/test_ivy/test_functional/test_experimental/test_nn/test_activations.py
@@ -1,3 +1,6 @@
+# global
+from hypothesis import strategies as st
+
 # local
 import ivy_tests.test_ivy.helpers as helpers
 from ivy_tests.test_ivy.helpers import handle_test
@@ -31,4 +34,38 @@ def test_logit(
         fn_name=fn_name,
         on_device=on_device,
         x=x[0],
+    )
+
+
+# thresholded_relu
+@handle_test(
+    fn_tree="functional.ivy.experimental.thresholded_relu",
+    dtype_and_x=helpers.dtype_and_values(
+        available_dtypes=helpers.get_dtypes("float"),
+        large_abs_safety_factor=8,
+        small_abs_safety_factor=8,
+        safety_factor_scale="log",
+    ),
+    threshold=st.floats(min_value=-0.10, max_value=10.0),
+)
+def test_thresholded_relu(
+    *,
+    dtype_and_x,
+    threshold,
+    test_flags,
+    backend_fw,
+    fn_name,
+    on_device,
+    ground_truth_backend,
+):
+    dtype, x = dtype_and_x
+    helpers.test_function(
+        ground_truth_backend=ground_truth_backend,
+        input_dtypes=dtype,
+        fw=backend_fw,
+        test_flags=test_flags,
+        fn_name=fn_name,
+        on_device=on_device,
+        x=x[0],
+        threshold=threshold,
     )
